be/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
mysql-connector-python==8.0.32
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
sentence-transformers==2.2.2
transformers==4.27.1
torch==2.0.0
bcrypt==4.0.1
python-decouple==3.8